simplify spmw tests
